@@ -15,10 +15,7 @@ macro_rules! stry {
 
 use std::io;
 use std::io::Write;
-use std::marker::PhantomData;
 use std::ptr;
-
-use crate::Value;
 
 const QU: u8 = b'"';
 const BS: u8 = b'\\';
@@ -81,7 +78,7 @@ pub trait BaseGenerator {
     /// Write a slice
     /// # Errors
     /// if the write fails
-    #[inline(always)]
+    #[inline]
     fn write(&mut self, slice: &[u8]) -> io::Result<()> {
         self.get_writer().write_all(slice)
     }
@@ -89,7 +86,7 @@ pub trait BaseGenerator {
     /// Write a char
     /// # Errors
     /// if the write fails
-    #[inline(always)]
+    #[inline]
     fn write_char(&mut self, ch: u8) -> io::Result<()> {
         self.get_writer().write_all(&[ch])
     }
@@ -102,45 +99,23 @@ pub trait BaseGenerator {
     /// writes new line
     /// # Errors
     /// if the write fails
-    #[inline(always)]
+    #[inline]
     fn new_line(&mut self) -> io::Result<()> {
         Ok(())
     }
 
     /// indents one step
-    #[inline(always)]
+    #[inline]
     fn indent(&mut self) {}
 
     /// dedents one step
-    #[inline(always)]
+    #[inline]
     fn dedent(&mut self) {}
-
-    /// Writes a string with escape sequences
-    /// # Errors
-    /// if the write fails
-    #[inline(never)]
-    fn write_string_complex(&mut self, string: &[u8], mut start: usize) -> io::Result<()> {
-        stry!(self.write(&string[..start]));
-
-        for (index, ch) in string.iter().enumerate().skip(start) {
-            let escape = ESCAPED[*ch as usize];
-            if escape > 0 {
-                stry!(self.write(&string[start..index]));
-                if escape == b'u' {
-                    stry!(u_encode(self.get_writer(), *ch));
-                } else {
-                    stry!(self.write(&[b'\\', escape]));
-                };
-                start = index + 1;
-            }
-        }
-        self.write(&string[start..])
-    }
 
     /// writes a string
     /// # Errors
     /// if the write fails
-    #[inline(always)]
+    #[inline]
     fn write_string(&mut self, string: &str) -> io::Result<()> {
         stry!(self.write_char(b'"'));
         stry!(self.write_string_content(string));
@@ -150,7 +125,7 @@ pub trait BaseGenerator {
     /// writes a string
     /// # Errors
     /// if the write fails
-    #[inline(always)]
+    #[inline]
     fn write_string_content(&mut self, string: &str) -> io::Result<()> {
         let mut string = string.as_bytes();
 
@@ -161,13 +136,8 @@ pub trait BaseGenerator {
             // this range.
             stry!(self.write_str_simd(&mut string));
         }
-        // Legacy code to handle the remainder of the code
-        for (index, ch) in string.iter().enumerate() {
-            if ESCAPED[*ch as usize] > 0 {
-                return self.write_string_complex(string, index);
-            }
-        }
-        self.write(string)
+
+        write_string_rust(self.get_writer(), &mut string)
     }
 
     /// writes a simple string (usually short and non escaped)
@@ -175,35 +145,28 @@ pub trait BaseGenerator {
     /// expensive on short strings.
     /// # Errors
     /// if the write fails
-    #[inline(always)]
+    #[inline]
     fn write_simple_string(&mut self, string: &str) -> io::Result<()> {
-        stry!(self.write_char(b'"'));
-
-        stry!(self.write_simple_str_content(string));
-        self.write_char(b'"')
+        self.write(br#"""#)?;
+        write_string_rust(self.get_writer(), &mut string.as_bytes())?;
+        self.write(br#"""#)
     }
-
     /// writes a simple string content  (usually short and non escaped)
     /// This means we can skip the simd accelerated writing which is
     /// expensive on short strings.
     /// # Errors
     /// if the write fails
-    #[inline(always)]
+    #[inline]
     fn write_simple_str_content(&mut self, string: &str) -> io::Result<()> {
-        let string = string.as_bytes();
+        let mut string = string.as_bytes();
         // Legacy code to handle the remainder of the code
-        for (index, ch) in string.iter().enumerate() {
-            if ESCAPED[*ch as usize] > 0 {
-                return self.write_string_complex(string, index);
-            }
-        }
-        self.write(string)
+        write_string_rust(self.get_writer(), &mut string)
     }
 
     /// writes a float value
     /// # Errors
     /// if the write fails
-    #[inline(always)]
+    #[inline]
     fn write_float(&mut self, num: f64) -> io::Result<()> {
         let mut buffer = ryu::Buffer::new();
         let s = buffer.format_finite(num);
@@ -213,144 +176,59 @@ pub trait BaseGenerator {
     /// writes an integer value
     /// # Errors
     /// if the write fails
-    #[inline(always)]
+    #[inline]
     fn write_int<I: itoa::Integer>(&mut self, num: I) -> io::Result<()> {
         let mut buffer = itoa::Buffer::new();
         let s = buffer.format(num);
         self.get_writer().write_all(s.as_bytes())
     }
 
-    #[cfg(target_feature = "avx2")]
-    #[inline(always)]
-    #[allow(clippy::cast_possible_wrap, clippy::cast_ptr_alignment)]
-    /// Writes a string with simd-acceleration
     /// # Safety
-    /// This function is unsafe because it uses simd instructions
+    /// This function is unsafe because it may use simd instructions
     /// # Errors
     ///  if the write fails
+    #[cfg(all(
+        feature = "runtime-detection",
+        any(target_arch = "x86_64", target_arch = "x86"),
+    ))]
     unsafe fn write_str_simd(&mut self, string: &mut &[u8]) -> io::Result<()> {
-        #[cfg(target_arch = "x86")]
-        use std::arch::x86::{
-            __m256i, _mm256_and_si256, _mm256_cmpeq_epi8, _mm256_loadu_si256, _mm256_movemask_epi8,
-            _mm256_or_si256, _mm256_set1_epi8, _mm256_xor_si256,
-        };
-        #[cfg(target_arch = "x86_64")]
-        use std::arch::x86_64::{
-            __m256i, _mm256_and_si256, _mm256_cmpeq_epi8, _mm256_loadu_si256, _mm256_movemask_epi8,
-            _mm256_or_si256, _mm256_set1_epi8, _mm256_xor_si256,
-        };
-
-        let writer = self.get_writer();
-        let mut idx = 0;
-        let zero = _mm256_set1_epi8(0);
-        let lower_quote_range = _mm256_set1_epi8(0x1F_i8);
-        let quote = _mm256_set1_epi8(b'"' as i8);
-        let backslash = _mm256_set1_epi8(b'\\' as i8);
-        while string.len() - idx >= 32 {
-            // Load 32 bytes of data;
-            let data: __m256i = _mm256_loadu_si256(string.as_ptr().add(idx).cast::<__m256i>());
-            // Test the data against being backslash and quote.
-            let bs_or_quote = _mm256_or_si256(
-                _mm256_cmpeq_epi8(data, backslash),
-                _mm256_cmpeq_epi8(data, quote),
-            );
-            // Now mask the data with the quote range (0x1F).
-            let in_quote_range = _mm256_and_si256(data, lower_quote_range);
-            // then test of the data is unchanged. aka: xor it with the
-            // Any field that was inside the quote range it will be zero
-            // now.
-            let is_unchanged = _mm256_xor_si256(data, in_quote_range);
-            let in_range = _mm256_cmpeq_epi8(is_unchanged, zero);
-            let quote_bits = _mm256_movemask_epi8(_mm256_or_si256(bs_or_quote, in_range));
-            if quote_bits == 0 {
-                idx += 32;
-            } else {
-                let quote_dist = quote_bits.trailing_zeros() as usize;
-                stry!(writer.write_all(string.get_unchecked(0..idx + quote_dist)));
-
-                let ch = string[idx + quote_dist];
-                match ESCAPED[ch as usize] {
-                    b'u' => stry!(u_encode(writer, ch)),
-                    escape => stry!(writer.write_all(&[b'\\', escape])),
-                };
-
-                *string = string.get_unchecked(idx + quote_dist + 1..);
-                idx = 0;
-            }
-        }
-        stry!(writer.write_all(&string[0..idx]));
-        *string = string.get_unchecked(idx..);
-        Ok(())
+        write_str_simd_fastest(self.get_writer(), string)
     }
-
-    #[cfg(all(not(target_feature = "avx2"), target_feature = "sse2"))]
-    #[inline(always)]
-    #[allow(clippy::cast_possible_wrap, clippy::cast_ptr_alignment)]
+    #[cfg(all(target_feature = "avx2", not(feature = "runtime-detection")))]
+    #[inline]
     /// Writes a string with simd-acceleration
     /// # Safety
     /// This function is unsafe because it uses simd instructions
     /// # Errors
     ///  if the write fails
     unsafe fn write_str_simd(&mut self, string: &mut &[u8]) -> io::Result<()> {
-        #[cfg(target_arch = "x86")]
-        use std::arch::x86::{
-            __m128i, _mm_and_si128, _mm_cmpeq_epi8, _mm_loadu_si128, _mm_movemask_epi8,
-            _mm_or_si128, _mm_set1_epi8, _mm_xor_si128,
-        };
-        #[cfg(target_arch = "x86_64")]
-        use std::arch::x86_64::{
-            __m128i, _mm_and_si128, _mm_cmpeq_epi8, _mm_loadu_si128, _mm_movemask_epi8,
-            _mm_or_si128, _mm_set1_epi8, _mm_xor_si128,
-        };
-
-        let writer = self.get_writer();
-        let mut idx = 0;
-        let zero = _mm_set1_epi8(0);
-        let lower_quote_range = _mm_set1_epi8(0x1F_i8);
-        let quote = _mm_set1_epi8(b'"' as i8);
-        let backslash = _mm_set1_epi8(b'\\' as i8);
-        while string.len() - idx > 16 {
-            // Load 16 bytes of data;
-            let data: __m128i = _mm_loadu_si128(string.as_ptr().add(idx).cast::<__m128i>());
-            // Test the data against being backslash and quote.
-            let bs_or_quote =
-                _mm_or_si128(_mm_cmpeq_epi8(data, backslash), _mm_cmpeq_epi8(data, quote));
-            // Now mask the data with the quote range (0x1F).
-            let in_quote_range = _mm_and_si128(data, lower_quote_range);
-            // then test of the data is unchanged. aka: xor it with the
-            // Any field that was inside the quote range it will be zero
-            // now.
-            let is_unchanged = _mm_xor_si128(data, in_quote_range);
-            let in_range = _mm_cmpeq_epi8(is_unchanged, zero);
-            let quote_bits = _mm_movemask_epi8(_mm_or_si128(bs_or_quote, in_range));
-            if quote_bits == 0 {
-                idx += 16;
-            } else {
-                let quote_dist = quote_bits.trailing_zeros() as usize;
-                stry!(writer.write_all(&string[0..idx + quote_dist]));
-
-                let ch = string[idx + quote_dist];
-                match ESCAPED[ch as usize] {
-                    b'u' => stry!(u_encode(writer, ch)),
-                    escape => stry!(writer.write_all(&[b'\\', escape])),
-                }
-
-                *string = &string[idx + quote_dist + 1..];
-                idx = 0;
-            }
-        }
-        stry!(writer.write_all(&string[0..idx]));
-        *string = &string[idx..];
-        Ok(())
+        write_str_simd_avx2(self.get_writer(), string)
     }
 
     #[cfg(all(
-        not(target_arch = "aarch64"),
-        not(all(target_arch = "wasm32", target_feature = "simd128")),
-        not(any(target_feature = "avx2", target_feature = "sse2")),
-        feature = "allow-non-simd"
+        target_feature = "sse2",
+        not(target_feature = "avx2"),
+        not(feature = "runtime-detection")
     ))]
-    #[inline(always)]
+    #[inline]
+    /// Writes a string with simd-acceleration
+    /// # Safety
+    /// This function is unsafe because it uses simd instructions
+    /// # Errors
+    ///  if the write fails
+    unsafe fn write_str_simd(&mut self, string: &mut &[u8]) -> io::Result<()> {
+        write_str_simd_sse42(self.get_writer(), string)
+    }
+
+    #[cfg(not(any(
+        feature = "runtime-detection",
+        feature = "portable",
+        target_feature = "avx2",
+        target_feature = "sse4.2",
+        target_feature = "simd128",
+        target_arch = "aarch64",
+    )))]
+    #[inline]
     /// Writes a string with simd-acceleration (not really, as the architecture doesn't support it)
     /// # Safety
     /// This function is unsafe because it uses simd instructions
@@ -360,17 +238,8 @@ pub trait BaseGenerator {
         self.write_simple_string(std::str::from_utf8_unchecked(string))
     }
 
-    #[cfg(all(
-        not(target_arch = "aarch64"),
-        not(all(target_arch = "wasm32", target_feature = "simd128")),
-        not(any(target_feature = "avx2", target_feature = "sse2")),
-        not(feature = "allow-non-simd")
-    ))]
-    unsafe fn write_str_simd(&mut self, string: &mut &[u8]) -> io::Result<()> {
-        compile_error!("write_str_simd not supported on the current architecture")
-    }
     #[cfg(target_arch = "aarch64")]
-    #[inline(always)]
+    #[inline]
     /// Writes a string with simd-acceleration
     /// # Safety
     /// This function is unsafe because it uses simd instructions
@@ -383,7 +252,7 @@ pub trait BaseGenerator {
         };
         use std::mem;
 
-        #[inline(always)]
+        #[inline]
         unsafe fn bit_mask() -> uint8x16_t {
             mem::transmute([
                 0x01_u8, 0x02, 0x4, 0x8, 0x10, 0x20, 0x40, 0x80, 0x01, 0x02, 0x4, 0x8, 0x10, 0x20,
@@ -391,7 +260,7 @@ pub trait BaseGenerator {
             ])
         }
 
-        #[inline(always)]
+        #[inline]
         unsafe fn neon_movemask(input: uint8x16_t) -> u16 {
             let simd_input: uint8x16_t = vandq_u8(input, bit_mask());
             let tmp: uint8x16_t = vpaddq_u8(simd_input, simd_input);
@@ -444,7 +313,7 @@ pub trait BaseGenerator {
     }
 
     #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
-    #[inline(always)]
+    #[inline]
     /// Writes a string with simd-acceleration
     /// # Safety
     /// This function is unsafe because it uses simd instructions
@@ -498,22 +367,290 @@ pub trait BaseGenerator {
     }
 }
 
+unsafe fn write_str_simd_fastest<W>(writer: &mut W, string: &mut &[u8]) -> io::Result<()>
+where
+    W: Write,
+{
+    // This is not possible right now since we can't get `W` to be part of the static
+    // use std::sync::atomic::{AtomicPtr, Ordering};
+    // type FnRaw = *mut ();
+
+    // type WriteStrFn<T> = for<'a, 'b, 'c> unsafe fn(&'a mut T, &'b mut &'c [u8]) -> io::Result<()>;
+    // static FN: AtomicPtr<()> = AtomicPtr::new(get_fastest::<W> as FnRaw);
+    // #[inline]
+    // fn get_fastest_available_implementation<W>() -> WriteStrFn<W>
+    // where
+    //     W: Write,
+    // {
+    //     if std::is_x86_feature_detected!("avx2") {
+    //         write_str_simd_avx2
+    //     } else if std::is_x86_feature_detected!("sse4.2") {
+    //         write_str_simd_sse42
+    //     } else {
+    //         write_str_simd_rust
+    //     }
+    // }
+
+    // #[inline]
+    // unsafe fn get_fastest<'invoke, 'de, W>(writer: &mut W, string: &mut &[u8]) -> io::Result<()>
+    // where
+    //     W: Write,
+    //     'de: 'invoke,
+    // {
+    //     let fun = get_fastest_available_implementation();
+    //     FN.store(fun as FnRaw, Ordering::Relaxed);
+    //     (fun)(writer, string)
+    // }
+    // let fun = FN.load(Ordering::Relaxed);
+    // mem::transmute::<FnRaw, WriteStrFn>(fun)(writer, string)
+
+    if std::is_x86_feature_detected!("avx2") {
+        write_str_simd_avx2(writer, string)
+    } else if std::is_x86_feature_detected!("sse4.2") {
+        write_str_simd_sse42(writer, string)
+    } else {
+        #[cfg(not(feature = "portable"))]
+        let r = write_string_rust(writer, string);
+        #[cfg(feature = "portable")]
+        let r = write_str_simd_portable(writer, string);
+
+        r
+    }
+}
+#[inline]
+fn write_string_complex<W>(writer: &mut W, string: &[u8], mut start: usize) -> io::Result<()>
+where
+    W: Write,
+{
+    stry!(writer.write_all(&string[..start]));
+
+    for (index, ch) in string.iter().enumerate().skip(start) {
+        let escape = ESCAPED[*ch as usize];
+        if escape > 0 {
+            stry!(writer.write_all(&string[start..index]));
+            if escape == b'u' {
+                stry!(u_encode(writer, *ch));
+            } else {
+                stry!(writer.write_all(&[b'\\', escape]));
+            };
+            start = index + 1;
+        }
+    }
+    writer.write_all(&string[start..])
+}
+
+#[inline]
+fn write_string_rust<W>(writer: &mut W, string: &mut &[u8]) -> io::Result<()>
+where
+    W: Write,
+{
+    // Legacy code to handle the remainder of the code
+    for (index, ch) in string.iter().enumerate() {
+        if ESCAPED[*ch as usize] > 0 {
+            return write_string_complex(writer, string, index);
+        }
+    }
+    writer.write_all(string)
+}
+
+#[cfg(feature = "portable")]
+#[inline]
+/// Writes a string with simd-acceleration
+/// # Safety
+/// This function is unsafe because it uses simd instructions
+/// # Errors
+///  if the write fails
+unsafe fn write_str_simd_portable<W>(writer: &mut W, string: &mut &[u8]) -> io::Result<()>
+where
+    W: Write,
+{
+    use std::simd::{u8x32, SimdPartialEq, ToBitMask};
+
+    let mut idx = 0;
+    let zero = u8x32::splat(0);
+    let lower_quote_range = u8x32::splat(0x1F_u8);
+    let quote = u8x32::splat(b'"');
+    let backslash = u8x32::splat(b'\\');
+    while string.len() - idx >= 32 {
+        // Load 32 bytes of data;
+        let data = u8x32::from_slice(&string[idx..]);
+        // Test the data against being backslash and quote.
+        let bs_or_quote = data.simd_eq(backslash) | data.simd_eq(quote);
+        // Now mask the data with the quote range (0x1F).
+        let in_quote_range = data & lower_quote_range;
+        // then test of the data is unchanged. aka: xor it with the
+        // Any field that was inside the quote range it will be zero
+        // now.
+        let is_unchanged = data ^ in_quote_range;
+        let in_range = is_unchanged.simd_eq(zero);
+        let quote_bits = (bs_or_quote | in_range).to_bitmask();
+        if quote_bits == 0 {
+            idx += 32;
+        } else {
+            let quote_dist = quote_bits.trailing_zeros() as usize;
+            stry!(writer.write_all(string.get_unchecked(0..idx + quote_dist)));
+
+            let ch = string[idx + quote_dist];
+            match ESCAPED[ch as usize] {
+                b'u' => stry!(u_encode(writer, ch)),
+                escape => stry!(writer.write_all(&[b'\\', escape])),
+            };
+
+            *string = string.get_unchecked(idx + quote_dist + 1..);
+            idx = 0;
+        }
+    }
+    stry!(writer.write_all(&string[0..idx]));
+    *string = string.get_unchecked(idx..);
+    Ok(())
+}
+
+#[target_feature(enable = "avx2")]
+#[inline]
+/// Writes a string with simd-acceleration
+/// # Safety
+/// This function is unsafe because it uses simd instructions
+/// # Errors
+///  if the write fails
+unsafe fn write_str_simd_avx2<W>(writer: &mut W, string: &mut &[u8]) -> io::Result<()>
+where
+    W: Write,
+{
+    #[cfg(target_arch = "x86")]
+    use std::arch::x86::{
+        __m256i, _mm256_and_si256, _mm256_cmpeq_epi8, _mm256_loadu_si256, _mm256_movemask_epi8,
+        _mm256_or_si256, _mm256_set1_epi8, _mm256_xor_si256,
+    };
+    #[cfg(target_arch = "x86_64")]
+    use std::arch::x86_64::{
+        __m256i, _mm256_and_si256, _mm256_cmpeq_epi8, _mm256_loadu_si256, _mm256_movemask_epi8,
+        _mm256_or_si256, _mm256_set1_epi8, _mm256_xor_si256,
+    };
+
+    let mut idx = 0;
+    let zero = _mm256_set1_epi8(0);
+    let lower_quote_range = _mm256_set1_epi8(0x1F_i8);
+    #[allow(clippy::cast_possible_wrap)] // it's a const, it's fine
+    let quote = _mm256_set1_epi8(b'"' as i8);
+    #[allow(clippy::cast_possible_wrap)] // it's a const, it's fine
+    let backslash = _mm256_set1_epi8(b'\\' as i8);
+    while string.len() - idx >= 32 {
+        // Load 32 bytes of data; _mm256_loadu_si256 does not require alignment
+        #[allow(clippy::cast_ptr_alignment)]
+        let data: __m256i = _mm256_loadu_si256(string.as_ptr().add(idx).cast::<__m256i>());
+        // Test the data against being backslash and quote.
+        let bs_or_quote = _mm256_or_si256(
+            _mm256_cmpeq_epi8(data, backslash),
+            _mm256_cmpeq_epi8(data, quote),
+        );
+        // Now mask the data with the quote range (0x1F).
+        let in_quote_range = _mm256_and_si256(data, lower_quote_range);
+        // then test of the data is unchanged. aka: xor it with the
+        // Any field that was inside the quote range it will be zero
+        // now.
+        let is_unchanged = _mm256_xor_si256(data, in_quote_range);
+        let in_range = _mm256_cmpeq_epi8(is_unchanged, zero);
+        let quote_bits = _mm256_movemask_epi8(_mm256_or_si256(bs_or_quote, in_range));
+        if quote_bits == 0 {
+            idx += 32;
+        } else {
+            let quote_dist = quote_bits.trailing_zeros() as usize;
+            stry!(writer.write_all(string.get_unchecked(0..idx + quote_dist)));
+
+            let ch = string[idx + quote_dist];
+            match ESCAPED[ch as usize] {
+                b'u' => stry!(u_encode(writer, ch)),
+                escape => stry!(writer.write_all(&[b'\\', escape])),
+            };
+
+            *string = string.get_unchecked(idx + quote_dist + 1..);
+            idx = 0;
+        }
+    }
+    stry!(writer.write_all(&string[0..idx]));
+    *string = string.get_unchecked(idx..);
+    Ok(())
+}
+
+#[target_feature(enable = "sse4.2")]
+#[inline]
+/// Writes a string with simd-acceleration
+/// # Safety
+/// This function is unsafe because it uses simd instructions
+/// # Errors
+///  if the write fails
+unsafe fn write_str_simd_sse42<W>(writer: &mut W, string: &mut &[u8]) -> io::Result<()>
+where
+    W: Write,
+{
+    #[cfg(target_arch = "x86")]
+    use std::arch::x86::{
+        __m128i, _mm_and_si128, _mm_cmpeq_epi8, _mm_loadu_si128, _mm_movemask_epi8, _mm_or_si128,
+        _mm_set1_epi8, _mm_xor_si128,
+    };
+    #[cfg(target_arch = "x86_64")]
+    use std::arch::x86_64::{
+        __m128i, _mm_and_si128, _mm_cmpeq_epi8, _mm_loadu_si128, _mm_movemask_epi8, _mm_or_si128,
+        _mm_set1_epi8, _mm_xor_si128,
+    };
+
+    let mut idx = 0;
+    let zero = _mm_set1_epi8(0);
+    let lower_quote_range = _mm_set1_epi8(0x1F_i8);
+    #[allow(clippy::cast_possible_wrap)] // it's a const, it's fine
+    let quote = _mm_set1_epi8(b'"' as i8);
+    #[allow(clippy::cast_possible_wrap)] // it's a const, it's fine
+    let backslash = _mm_set1_epi8(b'\\' as i8);
+    while string.len() - idx > 16 {
+        // Load 16 bytes of data; _mm_loadu_si128 does not require alignment
+        #[allow(clippy::cast_ptr_alignment)]
+        let data: __m128i = _mm_loadu_si128(string.as_ptr().add(idx).cast::<__m128i>());
+        // Test the data against being backslash and quote.
+        let bs_or_quote =
+            _mm_or_si128(_mm_cmpeq_epi8(data, backslash), _mm_cmpeq_epi8(data, quote));
+        // Now mask the data with the quote range (0x1F).
+        let in_quote_range = _mm_and_si128(data, lower_quote_range);
+        // then test of the data is unchanged. aka: xor it with the
+        // Any field that was inside the quote range it will be zero
+        // now.
+        let is_unchanged = _mm_xor_si128(data, in_quote_range);
+        let in_range = _mm_cmpeq_epi8(is_unchanged, zero);
+        let quote_bits = _mm_movemask_epi8(_mm_or_si128(bs_or_quote, in_range));
+        if quote_bits == 0 {
+            idx += 16;
+        } else {
+            let quote_dist = quote_bits.trailing_zeros() as usize;
+            stry!(writer.write_all(&string[0..idx + quote_dist]));
+
+            let ch = string[idx + quote_dist];
+            match ESCAPED[ch as usize] {
+                b'u' => stry!(u_encode(writer, ch)),
+                escape => stry!(writer.write_all(&[b'\\', escape])),
+            }
+
+            *string = &string[idx + quote_dist + 1..];
+            idx = 0;
+        }
+    }
+    stry!(writer.write_all(&string[0..idx]));
+    *string = &string[idx..];
+    Ok(())
+}
+
 ///  Simple dump Generator
-pub struct DumpGenerator<VT: Value> {
-    _value: PhantomData<VT>,
+pub struct DumpGenerator {
     code: Vec<u8>,
 }
 
-impl<VT: Value> Default for DumpGenerator<VT> {
+impl Default for DumpGenerator {
     fn default() -> Self {
         Self {
-            _value: PhantomData,
             code: Vec::with_capacity(1024),
         }
     }
 }
 
-impl<VT: Value> DumpGenerator<VT> {
+impl DumpGenerator {
     /// Creates a new generator
     #[must_use]
     pub fn new() -> Self {
@@ -529,26 +666,26 @@ impl<VT: Value> DumpGenerator<VT> {
     }
 }
 
-impl<VT: Value> BaseGenerator for DumpGenerator<VT> {
+impl BaseGenerator for DumpGenerator {
     type T = Vec<u8>;
 
-    #[inline(always)]
+    #[inline]
     fn write(&mut self, slice: &[u8]) -> io::Result<()> {
         extend_from_slice(&mut self.code, slice);
         Ok(())
     }
-    #[inline(always)]
+    #[inline]
     fn write_char(&mut self, ch: u8) -> io::Result<()> {
         self.code.push(ch);
         Ok(())
     }
 
-    #[inline(always)]
+    #[inline]
     fn get_writer(&mut self) -> &mut Vec<u8> {
         &mut self.code
     }
 
-    #[inline(always)]
+    #[inline]
     fn write_min(&mut self, _: &[u8], min: u8) -> io::Result<()> {
         self.code.push(min);
         Ok(())
@@ -556,14 +693,13 @@ impl<VT: Value> BaseGenerator for DumpGenerator<VT> {
 }
 
 /// Pretty Generator
-pub struct PrettyGenerator<V: Value> {
+pub struct PrettyGenerator {
     code: Vec<u8>,
     dent: u16,
     spaces_per_indent: u16,
-    _value: PhantomData<V>,
 }
 
-impl<V: Value> PrettyGenerator<V> {
+impl PrettyGenerator {
     /// Creates a new pretty printing generator
     #[must_use]
     pub fn new(spaces: u16) -> Self {
@@ -571,7 +707,6 @@ impl<V: Value> PrettyGenerator<V> {
             code: Vec::with_capacity(1024),
             dent: 0,
             spaces_per_indent: spaces,
-            _value: PhantomData,
         }
     }
 
@@ -582,26 +717,26 @@ impl<V: Value> PrettyGenerator<V> {
     }
 }
 
-impl<V: Value> BaseGenerator for PrettyGenerator<V> {
+impl BaseGenerator for PrettyGenerator {
     type T = Vec<u8>;
-    #[inline(always)]
+    #[inline]
     fn write(&mut self, slice: &[u8]) -> io::Result<()> {
         extend_from_slice(&mut self.code, slice);
         Ok(())
     }
 
-    #[inline(always)]
+    #[inline]
     fn write_char(&mut self, ch: u8) -> io::Result<()> {
         self.code.push(ch);
         Ok(())
     }
 
-    #[inline(always)]
+    #[inline]
     fn get_writer(&mut self) -> &mut Vec<u8> {
         &mut self.code
     }
 
-    #[inline(always)]
+    #[inline]
     fn write_min(&mut self, slice: &[u8], _: u8) -> io::Result<()> {
         extend_from_slice(&mut self.code, slice);
         Ok(())
@@ -626,38 +761,32 @@ impl<V: Value> BaseGenerator for PrettyGenerator<V> {
 }
 
 /// Writer Generator
-pub struct WriterGenerator<'w, W: 'w + Write, V: Value> {
+pub struct WriterGenerator<'w, W: 'w + Write> {
     writer: &'w mut W,
-    _value: PhantomData<V>,
 }
 
-impl<'w, W, V> WriterGenerator<'w, W, V>
+impl<'w, W> WriterGenerator<'w, W>
 where
     W: 'w + Write,
-    V: Value,
 {
     /// Creates a new generator
     pub fn new(writer: &'w mut W) -> Self {
-        WriterGenerator {
-            writer,
-            _value: PhantomData,
-        }
+        WriterGenerator { writer }
     }
 }
 
-impl<'w, W, V> BaseGenerator for WriterGenerator<'w, W, V>
+impl<'w, W> BaseGenerator for WriterGenerator<'w, W>
 where
     W: Write,
-    V: Value,
 {
     type T = W;
 
-    #[inline(always)]
+    #[inline]
     fn get_writer(&mut self) -> &mut W {
         self.writer
     }
 
-    #[inline(always)]
+    #[inline]
     fn write_min(&mut self, _: &[u8], min: u8) -> io::Result<()> {
         self.writer.write_all(&[min])
     }
@@ -665,21 +794,18 @@ where
 
 /// Pretty Writer Generator
 
-pub struct PrettyWriterGenerator<'w, W, V>
+pub struct PrettyWriterGenerator<'w, W>
 where
     W: 'w + Write,
-    V: Value,
 {
     writer: &'w mut W,
     dent: u16,
     spaces_per_indent: u16,
-    _value: PhantomData<V>,
 }
 
-impl<'w, W, V> PrettyWriterGenerator<'w, W, V>
+impl<'w, W> PrettyWriterGenerator<'w, W>
 where
     W: 'w + Write,
-    V: Value,
 {
     /// Creates a new generator
     pub fn new(writer: &'w mut W, spaces_per_indent: u16) -> Self {
@@ -687,24 +813,22 @@ where
             writer,
             dent: 0,
             spaces_per_indent,
-            _value: PhantomData,
         }
     }
 }
 
-impl<'w, W, V> BaseGenerator for PrettyWriterGenerator<'w, W, V>
+impl<'w, W> BaseGenerator for PrettyWriterGenerator<'w, W>
 where
     W: Write,
-    V: Value,
 {
     type T = W;
 
-    #[inline(always)]
+    #[inline]
     fn get_writer(&mut self) -> &mut W {
         self.writer
     }
 
-    #[inline(always)]
+    #[inline]
     fn write_min(&mut self, slice: &[u8], _: u8) -> io::Result<()> {
         self.writer.write_all(slice)
     }
@@ -730,8 +854,7 @@ where
 //
 // LLVM is not able to lower `Vec::extend_from_slice` into a memcpy, so this
 // helps eke out that last bit of performance.
-#[inline(always)]
-#[allow(clippy::uninit_vec)]
+#[inline]
 pub(crate) fn extend_from_slice(dst: &mut Vec<u8>, src: &[u8]) {
     let dst_len = dst.len();
     let src_len = src.len();
@@ -739,9 +862,8 @@ pub(crate) fn extend_from_slice(dst: &mut Vec<u8>, src: &[u8]) {
     dst.reserve(src_len);
 
     unsafe {
-        // We would have failed if `reserve` overflowed
-        dst.set_len(dst_len + src_len);
-
+        // We would have failed if `reserve` overflowed\
         ptr::copy_nonoverlapping(src.as_ptr(), dst.as_mut_ptr().add(dst_len), src_len);
+        dst.set_len(dst_len + src_len);
     }
 }
